refactor(frontend/controller): split up the search component
Split search component so that results are their own component
