+++ app/frontend/src/Controller/Search/index.css
